Make changes to prevent deadlocks and other minor improvements

+++ xv6-riscv/kernel/proc.h
@@ -108,7 +108,7 @@ struct proc {
 
 
   // ##################### Augmented as part of the Institutional assignments to test our understanding of the rich subject of operating systems #####################
-  uint64 creat_time;           // Time (in ticks) process p was created
-  uint64 start_time;           // Time (in ticks) process p was first scheduled to run 
-  uint64 end_time;             // Time (in ticks) process p was turned zombie in OS 
+  int creat_time;           // Time (in ticks) process p was created
+  int start_time;           // Time (in ticks) process p was first scheduled to run 
+  int end_time;             // Time (in ticks) process p was turned zombie in OS 
 };
